[nnbd_migration] Mark implicit dynamic for type params as nullable
Looks like we used to do this, but we stopped when we started changing
this pattern:

```
dynamic x = 1;
int y = x;
```

Our first iteration treated `dynamic` as nullable all the time. This is
not ideal in the example above. So we went back and instead began
treating `dynamic` as nullable only when it was observed to have a null
value.

However, in the case of unbounded type parameters:

```
class C<T> {}
```

We don't want to wait to observe:

```
C<Object> o = C(null);
```

before treating the bound as nullable.

Change-Id: I50b8ca6188d82cd62f795ada99a355afe2eae771
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/131714
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Mike Fairhurst <mfairhurst@google.com>