Create an abstract layer to customize querying collection elements.
This will be needed for the NNBD migration engine, so that when it
re-runs resolution to see what the types will be in migrated code, it
can tell the resolver to skip or modify collection elements to account
for dead code elimination.

The runtime overhead of this abstraction should be negligible, since
under normal analyzer operation, there will only be a single const
instance of this class, and no instances of any subclasses.  So the VM
should be able to trivially inline all uses of the abstraction layer.

Also add SetOrMapLiteralImpl.becomeUnresolved(), which is used to
reset literals to the "unresolved" state before they are resolved.
This will be necessary so that when the migration engine re-resolves a
set or map literal, the determination of what kind of literal it is
will be made afresh.

Change-Id: Id8e95c944acb892c60d3d92a7facc20d08604211
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/131842
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>