Add some minimal presubmit checks to keep forked SDK tests in sync.
This only validates that if a test and its fork exist that the CL
touches both. There are lots of other cases we could imagine checking
like if you *add* a new test on one side, you should one on the other.
But those are likely to have a lot of false positives now while the
migration is happening and new NNBD tests are being written.

So for now, this just tries to check that tests that have been forked
at least don't diverge. It's a start.

Change-Id: Id6a9c39699a510da1a4a4702879210f94cd69c26
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/131945
Auto-Submit: Bob Nystrom <rnystrom@google.com>
Reviewed-by: Jonas Termansen <sortie@google.com>
Commit-Queue: Bob Nystrom <rnystrom@google.com>