Dartfix: add fixes, cleanup output, fix bug
This adds support for running the remaining lint fixes that were already
implemented but not yet supported in dartfix.

It cleans up the output by shortening the descriptions of most of the
pre-existing fixes.

It also marks the recently added fixes as not being default, because
making them default was unintentional and should be considered to be
a bug.

Change-Id: I381924144c89a8c8666d6f16009724ae0f2e1fbc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/132540
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>