Update the nnbd-mode flag name according to the proposed specification.
See: https://github.com/dart-lang/language/pull/779/files

Currently this only handles one or the other configuration via the flag.
We might have to add extra logic to change the default when no flag is provided
depending on how that proposal lands (likely the requirement will be to default
to weak if the main entry point is opted-out, but default to strong otherwise).

Change-Id: Iebf573f251de6a3a9cf93781b76061f41c181f3c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/131761
Reviewed-by: Mayank Patke <fishythefish@google.com>