Migration: as a temporary measure, remove code via commenting it out.
The new FixBuilder infrastructure has the ability to remove dead code
either by commenting it out or by deleting it entirely.  Previous to
FixBuilder all we could do is comment it out.  My intention when
landing FixBuilder was to keep this behavior so that we could make a
deliberate switch at a later time (when everything was well tested).
But I accidentally set the flag wrong for some code paths, and so
currently we delete code.

This has caused some bugs in the preview tool (which I will file in
the issue tracker shortly).  In order to avoid causing problems for
members of the language team who intend to try out the tool this week,
I'm switching back to "comment code out" as a temporary measure.

Change-Id: I0ca8674b6a8d9ab23cf6726a15fa3dc1919085b4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/133081
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>