[io/file_watcher] Unregister watched pathId when handle is closed.
When file watching stream reports an error, it will close the handle too after sending "closed" event out to the listeners. Listener should not attempt to do anything with that handle(kept as pathId in the listener). In particular it should not attempt to call that handle to unwatch when user of the listener requested to cancel the listener.

Further fixes for https://github.com/flutter/flutter/issues/46616

Change-Id: Iccc100e294b11d3999ad68e775ef1ceb1d7a5efb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/133180
Commit-Queue: Alexander Aprelev <aam@google.com>
Reviewed-by: Zichang Guo <zichangguo@google.com>