Migration: allow a node to become nullable after being marked non-nullable.
Previously, graph propagation occurred in three steps:

1. Mark any nodes connected to `always` by union edges as nullable.

2. Mark any nodes connected to `never` by hard or union edges as
   non-nullable, but do not change any nodes marked by #1.

3. Mark any nodes connected to `always` as nullable, but do not change
   any nodes marked by #1 or #2 (with additional processing for
   substitution nodes).

This approach had the advantage that once a node is marked as nullable
or non-nullable, it is never changed, which made reasoning about the
algorithm a bit easier.  But in an upcoming CL I'll need to modify the
algorithm in a way that's incompatible with this approach.

This CL introduces the notion that some edges are "checkable" and some
aren't (an edge is "checkable" if it's ok for graph propagation to
leave it unsatisfied, because the fix builder can insert a null
check).  At the moment all edges are considered checkable except for
union edges, but that will change in future CLs in order to address
https://github.com/dart-lang/sdk/issues/38699.

The algorithm is now:

1. Mark any nodes connected to `never` by hard or union edges as
   non-nullable.

2. Mark any nodes connected to `always` as nullable, but only change
   nodes marked by #1 if necessary to satisfy uncheckable edges (with
   additional processing for substitution nodes).

An alternative way of thinking of this algorithm is that hard edges
reflect user intentions that we should try to satisfy if we can, but
which shouldn't take priority over creating output that is free of
static errors.  During phase 1, we propagate hard edges to find all
the nodes the user *intends* to make non-nullable.  During phase 2, we
propagate nullability, respecting user intention where we can, and
overriding it where we can't.

Change-Id: Ie3f9d9ee66f407f9d4916f7259e7f50a82ebb2bb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/133402
Reviewed-by: Mike Fairhurst <mfairhurst@google.com>