Migration: consider edges introduced by contravariant assignment to be uncheckable.
This prevents a non-null assertion in a callback function from marking
the parameter type as non-nullable, if that would make the callback
function fail to meet the requirements of its use site.  For example,
when migrating:

  f(List<int/*?*/> myList) => myList.map((int x) => x + 1);

We previously would produce:

  f(List<int?> myList) => myList.map((int x) => x + 1);

(Which would fail immediately due to the callback `(int x) => x + 1`
not accepting `int?`).  We now produce:

  f(List<int?> myList) => myList.map((int? x) => x! + 1);

(Which only fails if there is actually a null value in the list,
matching the behavior of the code prior to migration).

Addresses another manifestation of
https://github.com/dart-lang/sdk/issues/38699.

Change-Id: Iffe9fbdace00ff1edea0f6d125c86b06fad7bbcb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/133423
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Mike Fairhurst <mfairhurst@google.com>