[CFE] Don't add newly-converted DillLibraryBuilders to list of needed dill libraries
When d21c4e4fa5ef8ab7e8dc1423f15c2fb1966be3e8 ([CFE] [Incremental
compiler] Convert all SourceLibraryBuilders to DillLibraryBuilders)
landed it also potentially (and inadvertently) changed the "needed dill
libraries" which tracks which libraries - from a dill file - was
actually needed to do that compilation.

Before, only libraries actually coming from a dill file was in this
list.
After, if asking the class hierarchy for information on the
newly-compiled classes, and then asked for the list of needed dill
libraries to be updated (the normal case in DDC) the libraries that was
just compiled (and converted to dill libarries) was added to the list.
DDC in turn tried to attach all marked libararies to the specific
(actual, "physical") dill file it was loaded from, but as those libraries
wasn't loaded from any dill "physical" dill file, it would crash.

This CL fixes the issue by remembering that these are special.

Change-Id: I037ee8f653165d94b5904e4beacc2423801a3461
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/133523
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>